backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
pydantic==1.10.13
PyJWT==2.8.0
python-multipart==0.0.6
websockets==11.0.3